+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "form-render",
-  "version": "0.4.3-beta3",
+  "version": "0.4.3-beta5",
   "description": "通过 JSON Schema 生成标准 Form，常用于自定义搭建配置界面生成",
   "repository": {
     "type": "git",
